Hopefully improved the UI performance: Changed BulkWrite to do indeed a BulkWrite of a number of pixels from an array (used in WaterfallDisplay) And replaced almost all SPI transfer with the more optimized *Fast.
Please test carefully. In SPI mode I can use O3 w/ no problems.
Cannot test parallel with O3. My hunch is that the SRAM timing is now
too fast and should be reconfigured to be slow since there are less
instructions inbetween calls and the display controller may not be
able to keep up with the speed of the MCU unless given some break.
Should be fixable by changing the SRAM interface timing.