Firmware Change Related Issue: Disabled some potentially "dangerous" code. Dangerous means it may rather create problems than solve the one it is intended to solve. Background: The original code detected firmware change and more or less immediately save the current settings including the new firmware version to EEPROM / Flash, leaving the user no choice. Intention was to update the configuration settings. Good idea. However, before this the settings have been read from EEPROM. Already at this point they have been "clean", i.e. boundary checked and if necessary reset to proper values, so mcHF will operate correctly anyway.
And the mandatory saving of the enforced changes as implemented
in the old code is counter-productive in testing of new firmwares.
If the user switches back and force for testing, both firmware try to
sanitize the values, possibly resulting in changed settings or loss
of settings. Now user can switch off without saving and changes to
configuration are not recorded unless the user wants it.

In normal operation all changes get saved when turning off with the
power button anyway. This will save the changes permanently.
so no harm is done.