Reworked RX-TX Switching code.
Some things did not happen in the right order,
so that an visible (and possibly) audible strong noise
was generate when switching to TX. Also caused the first
morse dit or dah to be distorted.

Right now a delay of 100ms exists before TX starts in
FM and AM after closing the PTT switch. As far as I can see,
the noise is gone.

CW has less latency in most cases, DAH starts roughly
after 30ms, but here some initial unwanted noise at the beginning
is still there.
However, it is not the switching noise, it is a very short
start of a morse dit/dah, which is restarted after around 10ms.

Needs further investigation how to get rid of this.

NEEDS TESTING ESPECIALLY IN CW MODE.