Securing CAT frame sizes to 5 byte packets
In the situation where USB buffers are filled with numbers other than
multiples of five, it is likely that the bytes have been added in front.
The bytes in the beginning of the buffers are probably left-ovres from
previous sessions or non CAT programs acessing the serial port.
By this fix the first bytes of the USB read buffer is flushed as long as
the rest of the buffer-size divided by five is not zero. The
calcuilation time is normally minimal.