All voice tx modes now use the same input processing filter(s).
Slightly restructured tx processing in order to simplify,
no functional changes.

Used more "const" and local copies of volatile variables to see the effect.
Codewise we lost a few hundreds byte, effect on performance
is not so easy to measure, but for sure not slower.

Removed volatile modifier from spectrum display data, since it is not necessary IMHO
here.