Implemented  "Keep Signal Option" for CW Shift Offset Mode What it does: Best of both worlds (in a way) Like in RX or TX, if you are tuned to a signal around the CW sidetone frequency in SSB you will stay tuned to this signal if going into CW mode. I.e. if your sidetone is 700 Hz and you are tuned to 7.000.000 Hz and you hear a CW signal with about 700 Hz, simple switch to CW. You will still hear the signal and your frequency display will show 7.000.700 Hz, since this will be your TX frequency.
Unlike CW RX or TX offset modes it works even if you are switching sidebands, e.g. from
SSB LSB to CW-U (USB). Also if you switch between CW-U and CW-L you'll stay
tuned to the signal (as it was in CW Shift Mode). If you switch back to SSB,
you will still be tuned to that signal.

How to operate: In CW menu, set CW Offset to "Shift". In Debug menu, switch on "CW Shift Keeps Signal".
Tune to a CW signal around the sidetone frequency, and switch to CW.