Factored frequency shift code of of audio_driver.c Added high performance shift algorithm for arbitrary shifts based on Clint's (KA7OEI) hint. Works very well. Please note, it is not "active" at the moment since we only use fixed shift frequencies which can be handled with the previously used algorithms. It becomes automatically active if the sample frequency is not a power of 2 multiple of the shift frequency
Also "simplified" the algorithms somewhat by exploiting the shift direction
reversal if I and Q are exchanged as inputs into the algorithm.

Tuned the use of double math in strategic places. This was tested on a
PC and final performance validated on a STM32F4 machine. Code is as fast as before.

Tried to comment extensively.