Enabled 32bit ADC/DAC sample transfers
We now run the codec(s) with their full data width (24bits)
(transfer size 32bit I2S data). Purpose is to have a little more margin
if your frontend does not deliver high enough IQ levels. It may or may
not improve the reception. It has almost no effect on performance since
we, as before, run most of the processing with floating point math.
A little extra work is on the DMA and SAI hardware (transfer amount doubled.
And a little conversion is necessary for now.

On platforms with 2 codecs you can select AUDIO and IQ bit width (16/32)
in any combination, on single codec devices (mcHF)
both have to be identical.

On STM32F4 we have an ugly hack to fix our inability to get proper 32bit via
DMA. The current approach is required for a working solution
as we have issues when working on the DMA buffers directly for swapping (we have to fix the
MPU config for that to work, I guess)

RAM usage grows by 0.5 (mcHF) to 1k (OVI40).
USB I/O still uses only 16bits (also for IQ data).
Audio driver load does not increase measurably.