Update Eclipse settings to match Makefile, fixed issue fixed library list (no more libc, libm or libnosys explicitly mentioned) Add nosys.specs
Fixed a failing assert in the HAL RCC driver in the OVI40 builds.
It was triggered by a PLL value check on a PLL output (PLLR) we are not using.
It never showed up before due to the compiler optimization.
Now with the endless loop in assert_failed, the compiler figured, no more code
should be executed (hence the binary was very small now).

Unfortunately, USE_FULL_ASSERT can no longer used on the OVI40 bootloader build,
it won't fit then so it was removed from the DebugBootloaderOVI40 config.