PA Level Control refactoring - tried to separate the PA power steps (which are more a UI thing) from the calculation of the settings to control power output by the PA. This makes it easier to connect different PA (with differing power levels) using the same code (which is at the moment basically the calculation of a gain value to interface the DSP output with the PA side of things). - Also changed display code to dynamically calculate the power display. Range 0mW to 100W will fit into existing box.
I also added color coding by background color of the nominal power output box:

Blue -> normal

Orange -> the effective power setting is lower than the requested value,
i.e. when switching to AM mode, we limit it to 2W, and now user can see this.
Also the reduced out-of-band TX power is shown this way. And we now have
a better (not good!) control here, as we limit the power out of band to 5/50mW.
This power estimation is poor but better than the fixed gain value before.

Red -> there is an problem in the calibration, output gain is 0,
i.e. no output is produced, this happens for uncalibrated machines.

Gray -> TX is disabled, hence no power output.

Essentially, anything not with blue background means you'll get
less or no output power than requested.