Fixed CW input for RTTY and BPSK (part 2)
- RTTY and BPSK now accept input from USB keyboard AND morse paddles (if
  an keyer mode is set, not for Straight Key).

Note: I found in cw_gen.c the beginning of a keyboard to morse converter.
This was eating RTTY/BPSK keyboard input. It seems not to be operational
since it requires the TRX in TX mode to accept input, but in CW the starts
only once you use a paddle...