Added support for LAPWING Hardware
This code is not complete, the formula in Si570_translateExt2Osc is not correct.
It also does not disable the Si5351A, which is not available on the LAPWING.

And last but not least, it has not been tested on the Lapwing at all.
Travis shows how to create a Lapwing Binary