Memory Mode implemented
Memory Mode differs from the default band mode in the use of memories.
In band mode each memory is linked to a specific ham band. Only memories
which are linked to a supported band will be enabled and can be reached via
Band+/-. In Memory Mode, all memories are enabled and can store any frequency.
In Band mode, a memory name is Band<Wavelength>, in Memory mode it is Mem<MemNum>

Currently this is a compile time switch, but it would trival to convert to
a configuration switch.

By default this is only enabled for the LAPWING as this is a mono band trx
which can make use of this mode nicely.