Fix digit flickering (experimental)
The digit flickering for the main display should be gone now.
However, this is only a partial fix. Conceptually the flickering
still happens except for the main display using the 8bit font.
It is not noticable for the small frequency display and no build currently
uses a non-8bit font for the main display.

It is still a kind of a mess until we support variable width fonts also for
the 1bit fonts or we have to draw the dots "by hand" between digits groups.