Fix for #1891 FT-817 CAT emulation broken for WINTEST
The change back to the original response of the real FT-817
on PTT_STATE CAT command (which is 0xff) fixes an issue with the
WINTEST program.

That hamlib only tests the bit 7 (0x80) in ft817_get_ptt does not make 0x80 a valid
answer for other programs.

cherry-pick cat fix